int str float oh my

+++ joybot.py
@@ -27,7 +27,7 @@ while True:
    c.send('Thank you for connecting')
    joydata = c.recv(1024).decode()
    print (joydata)
-   joydata = int(joydata)
-   c.send('received joydata', joydata)
+   joydata = float(joydata)
+   c.send('received joydata', (joydata))
    #c.close()                # Close the connection
    #some changes to commit
